app/requirements.txt
fastapi
uvicorn
numpy
pillow
requests
python-multipart